+++ src/components/ForestTypeModal/ForestTypeComparison/ComparisonCell.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import ForestTypeLink from './ForestTypeLink';
+import ForestTypeLink from '../ForestTypeLink';
 import comparisonStyles from './ForestTypeComparison.module.css';
-import { getStringWithUnit } from '../../utils/comparisonUtils';
-import useIsMobile from '../../hooks/useIsMobile';
+import { getStringWithUnit } from '../../../utils/comparisonUtils';
+import useIsMobile from '../../../hooks/useIsMobile';
 
 const ComparisonCell = ({
   data,
@@ -17,11 +17,11 @@ const ComparisonCell = ({
 }) => {
   const isMobile = useIsMobile();
   return (
-    <td className={className || undefined} style={{ verticalAlign: 'top' }}>
+    <td className={className || ''} style={{ verticalAlign: 'top' }}>
       <>
         <span style={{ display: 'flex' }}>
           {isMobile && (
-            <span style={{ minWidth: 50 }}>
+            <span style={{ minWidth: 60 }}>
               <ForestTypeLink code={code} />:
             </span>
           )}
