+++ src/components/ForestTypeModal/ForestTypeComparison/ForestTypeComparison.module.css
