+++ src/components/ForestTypeModal/ForestTypeComparison/bl/index.js
