+++ src/components/ForestTypeModal/ForestTypeComparison/ch/index.js
