+++ src/components/ForestTypeModal/ForestTypeComparison/lu/index.js
