+++ src/components/ForestTypeModal/ForestTypeDescription/Diagram.module.css
