+++ src/components/ForestTypeModal/ForestTypeDescription/ForestTypeDescription.module.css
@@ -10,6 +10,7 @@
   text-decoration: underline;
   cursor: pointer;
   text-align: left;
+  margin-bottom: 5px;
 }
 
 .link:hover {
