+++ src/components/ForestTypeModal/ForestTypeDescription/Legend.js
