+++ src/components/ForestTypeModal/ForestTypeDescription/ch/AltitudinalZoneForestEcoregion.js
