+++ src/components/ForestTypeModal/ForestTypeDescription/ch/ForestTypesDescription.js
