+++ src/components/ForestTypeModal/ForestTypeDescription/ch/Grid.js
