+++ src/components/ForestTypeModal/ForestTypeDescription/ch/Humus.js
