+++ src/components/ForestTypeModal/ForestTypeDescription/ch/SkeletalFraction.js
