+++ src/components/ForestTypeModal/ForestTypeDescription/ch/SoilDiagram.js
