+++ src/components/ForestTypeModal/ForestTypeDescription/ch/SoilTab.js
