+++ src/components/ForestTypeModal/ForestTypeDescription/ch/TerrainTab.js
@@ -6,7 +6,7 @@ import { Table } from 'semantic-ui-react';
 import AltitudinalZone from './AltitudinalZoneForestEcoregion';
 import Legend from '../Legend';
 import Site from './Site';
-import useIsMobile from '../../../hooks/useIsMobile';
+import useIsMobile from '../../../../hooks/useIsMobile';
 
 function TerrainTab({ data }) {
   const isMobile = useIsMobile();
