+++ src/components/ForestTypeModal/ForestTypeDescription/ch/VegetationTab.js
