+++ src/components/ForestTypeModal/ForestTypeDescription/index.js
