+++ src/components/ForestTypeModal/ForestTypeDescription/utils.js
