+++ src/components/ForestTypeModal/index.js
@@ -1,0 +1,1 @@
+export { default } from './ForestTypeModal';
