+++ src/i18n/resources/fr/translation.json
@@ -201,6 +201,7 @@
     "descriptionNaturalForest": "Description forêt naturelle",
     "descriptionForestType": "Description emplacement",
     "vegetationIndicator": "Groupes d'indicateurs",
+    "geology": "Géologie",
     "aspect": {
       "001-025": "Nord à nord-nord-est",
       "026-050": "Nord-nord-est à nord-est",
